#CLOUD-511 added 'range' match algorithm support to stree condition node
#CLOUD-511 cloud module endpoint is now defined by stree xml. Needed to be able to use different cloud instances depending on various parameters
#CLOUD-511 printing to log done right
#CLOUD-511 some fixes
#VMS-2359 added check for offline status
#VMS-2388 chunked encoding fix
#VMS-2454 multiple credentials support for axis searcher
#VMS-2454 possible fix
#VMS-2774 DW CPRO h264 profile and advanced settings page fixes
#VMS-2980 fixed speech synthesizer output sample rate
#VMS-3050 fixed possible server-to-server connection lock
#VMS-3050 peer id presence in existing transaction connection container is now controlled with a guard
#VMS-3050 review fixes
#VMS-3176 integration in progress
#VMS-3325 #VMS-3328 #VMS-3330 #VMS-3326 fixes
#VMS-3325 #VMS-3328 #VMS-3330 #VMS-3326 fixes, in progress.
#VMS-3328 in progress
#VMS-3328 wrong IO ports merge fix
#VMS-3330 Auto reset timeouts fix
#VMS-3331 fixes. Added resource status monitoring & network issue handler
#VMS-3360 MD maximum resolution increased
#VMS-3416 infinite waiting on wait condition in QueueFileWriter fixed
#VMS-3447 IR cut filter advanced parameter fix
#VMS-3475 possible fix
#VMS-3475 stupid error fix
#VMS-3478 onvif reader improvements
#vms-2287. Logging opened file handles.
#vms-2287. merge vms_2287 --> dev_2.6.
#vms-2618. Graceful windows service restart.
#vms-3141. Adaptive write buffer size. MediaCycleBuffer assert fix.
#vms-3141. Adaptive write buffer size. More logs. Max ffmpeg buffer size setting added.
#vms-3318. Buffered file && cycle buffer fixes.
#vms-3536. Mac build fix.
#vms-3536. Mac build fix. Merge heads.
#vms-3536. Mac build fix. Merge prod_3.0.0 --> dev_3.0.0
#vms-3551. Backed out some changes.
(VMS-2576) - 2way-audio button disables 'Aero'
(VMS-3078) - Setup Wizard Advanced Server settings checkboxes do not change flag state in api/systemSettings
(VMS-3141) - Make buffer size setting automatical and adaptive
(VMS-3141) Adaptive ffmpeg buffer limits collapsed to single value
(VMS-3400) - AV2195 and AV2196 added to resources xml
(VMS-3401) - scripts for new av and axis cameras search are updated
(VMS-3442) - Do not run setup wizard if the system is in Safe Mode
(VMS-3471) - review minor
(VMS-3499) - Irresponsable QnConcurrent::run crash in QnPlAxisResource
(VMS-3500) - Time lapse: video exported with timestamps has wrong rapid review speed
(VMS-3512) - Handle merging two cloud systems
2.6 Development branch
2.6 Production branch
ADAM-6000 initInternal improvement
Added auto discovery and support for 6050, 6051, 6060, 6066 modules
Added tag before_ipv6 for changeset 121495748bff
Advantech ADAM module integration and Modbus client. In progress.
Automatic dialog size fix.
Autoreset timeout fix
Backed out changeset c5c790a9cad4
Backed out changeset: f90e3c76cf18
Branch for fix
Busy indicator in the login to cloud dialog.
CLOUD-499 - Automatically login to Cloud portal from NxWitness, if server is connected to cloud
Compile fix
Default cloud portal links are HTTPS now
Fix compile error
Fix iOS compile error
Flir stream reader enhancements
Left-aligned push buttons with customizable margin.
Linux compile fix
Linux compile fix.
Log all web page errors in the client
Login to cloud dialog UI tweaks
Login to cloud dialog UI tweaks.
MOBILE-212: Disable screen orientation locking.
MOBILE-245 IP version 6 support
MOBILE-245 IPv6 review fixes
MOBILE-251: Increase timeout for video loading.
MOBILE: Adjust camera info animation.
MOBILE: Change mouse behavior in lite client.
MOBILE: Disable screen orientation locking.
MOBILE: Do not quit app by escape.
MOBILE: Fix initial display of offline/unauthorized cameras.
MOBILE: Lite client help popup.
Manual searcher fixes
Minor fix
Minor fix.
Minor syntax sugar
Minor visual bug fix
Mobile client build fixes
Move Qn::ItemDataRole to client lib
Oops.
PngCrush all png files
QnSkin::maximumSize() slightly changed. QnSkin::maximumSizePixmap() added.
Refactoring & UI tweaks.
Remove unneeded include
Review fix
Review fixes.
Style fix
Swapped icons fix
Temporary fix timeline behavior
Update rect fix.
Useless dw stream reader was removed
VISTA-89 address review comments
VISTA-89 e-mail on a locally based email server
VMS-2523 - SMTP servers without authentication aren't supported
VMS-2602: Server Settings Redesign: Storage Analytics
VMS-2602: Server Settings dialogs redesign - Storage Management
VMS-2608: Login to cloud dialog redesign
VMS-2608: Redesign cloud dialogs
VMS-2795 - Redesign title bar, add offline state to cloud panel
VMS-2874: Add URL protocol association to mimeapps.list
VMS-3061 - Need to optimize User Settings -> 'Media Resources' for big data
VMS-3104: Switch client to local mode when some files has been dragged onto welcome screen
VMS-3111: Minilauncher for MacOS
VMS-3176 added debounce for io monitor
VMS-3297: Update applauncher desktop file.
VMS-3314 - Cannot start NX Server Tool
VMS-3326 - Increase width of column '#' on 'I/O Ports' form
VMS-3327 - Disable 'Recording' tab for ADAM-6052
VMS-3332 - Asterisk for modified shared layouts must be present in all Tree nodes
VMS-3387: Make QnMediaServerResource::getApiUrl return https for ssl-enabled servers.
VMS-3398: Add information about software version on welcome screen
VMS-3402: Welcome screen with no systems
VMS-3404 - Client should be opened and logged in to system after 'Open in NX Witness' is pressed
VMS-3409 - Make sure launcher works to be able to export *.exe and view it
VMS-3413 - Assign hotkey for 'Logout action'
VMS-3425
VMS-3435 - Exported layouts are not shown in tree in offline mode
VMS-3469: Show alert if Client 2.6.0 is opened on hiDPI screen
VMS-3483 - null pointer fix
VMS-3486 - Ability to set default background parameters for customizations
VMS-3503 - Wrong 'tree' list in 'Thick' client after Unlink System from Cloud
VMS-3508 - Can't set local smtp server at 'System Administration'
VMS-3510 fix possible io-module client freeze
VMS-3510 review fix
VMS-3510 review fix 2
VMS-3532: Incorrect cox and tricom .dmg
VMS-3539 - Dragging of custom user local layout to another user removes it from custom user layouts tree list
VMS-3540 - deleting of shared Layout at 'tree' list from custom user doesn't remove it from User Settings window Layouts tab
VMS-3543 - Web page can't be added to users local layout by dragging into it at 'tree' list
VMS-3546 - Instantly save all changes with cameras and layouts made in resources tree
VMS-3555: MacOS: Do not allow to connect to server with version less than 3.0
VMS-3557: MacOS: Client crashes on restart after update
VMS-3557: MacOS: Client crashes on restart after update: compilation fix try
VMS-3573 - Server on Windows stopped working with its config after update
VMS: Fix a gcc warning.
VMS: Fix auto_ptr warnings.
Version bump
added forceSingleStream parameter
added missing #include <>
compile error fixes
compile fix
compile fixes
cox - icon size corrected
filetypes.py removed
fix
fix nx_networ_ut compilation
fix some annoying warnings
fixed bug after merge
fixed compile error
fixed compile error after mere
integration in progress
ionetworks - mac installer background
ionetworks - mac installer background adjusted
ionetworks - removing unnecessary files
ionetworks - user manual
ionetworks customization
ionetworks customization - adjustments
ionetworks customization - corrections
ionetworks eula - linux
ionetworks eula - windows
ipv6 fixes for WIN32
last test fix after merge
linux compile fix
little fix
nx network post merge compile fixes
nx::network::ExtendedStreamSocket - socket wrapper with extended functionality
nx_network_ut runtime fixes
review fixes
some possibly lost calls of av_init_packet
some post merge compile fixes
win32 - ipv6 fixes
wrong semicolon fix

<==== Mercurial data ====>
Branch: x_vms_3514_merge_update

+++ client/libclient/src/camera/cam_display.cpp
@@ -308,7 +308,7 @@ void QnCamDisplay::hurryUpCkeckForCamera2(QnAbstractMediaDataPtr media)
         //bool isPrebuffer = media->flags & QnAbstractMediaData::MediaFlags_FCZ;
         if (m_speed < 1.0 || m_singleShotMode)
             return;
-        if ((quint64)m_firstAfterJumpTime == AV_NOPTS_VALUE) {
+        if (m_firstAfterJumpTime == AV_NOPTS_VALUE) {
             m_firstAfterJumpTime = media->timestamp;
             m_receivedInterval = 0;
             m_afterJumpTimer.restart();
@@ -530,7 +530,7 @@ bool QnCamDisplay::display(QnCompressedVideoDataPtr vd, bool sleep, float speed)
         {
             qint64 displayedTime = getCurrentTime();
             //bool isSingleShot = vd->flags & QnAbstractMediaData::MediaFlags_SingleShot;
-            if (speed != 0  && (quint64)displayedTime != AV_NOPTS_VALUE && m_lastFrameDisplayed == QnVideoStreamDisplay::Status_Displayed &&
+            if (speed != 0  && displayedTime != AV_NOPTS_VALUE && m_lastFrameDisplayed == QnVideoStreamDisplay::Status_Displayed &&
                 !(vd->flags & QnAbstractMediaData::MediaFlags_BOF))
             {
                 NX_ASSERT(!(vd->flags & QnAbstractMediaData::MediaFlags_Ignore));
@@ -595,7 +595,7 @@ bool QnCamDisplay::display(QnCompressedVideoDataPtr vd, bool sleep, float speed)
                 realSleepTime = m_delay.addQuant(needToSleep);
         }
         //qDebug() << "sleep time: " << needToSleep/1000.0 << "  real:" << realSleepTime/1000.0;
-        if ((quint64)realSleepTime != AV_NOPTS_VALUE && !m_buffering)
+        if (realSleepTime != AV_NOPTS_VALUE && !m_buffering)
             hurryUpCheck(vd, speed, needToSleep, realSleepTime);
     }
 
@@ -708,7 +708,7 @@ bool QnCamDisplay::doDelayForAudio(QnConstCompressedAudioDataPtr ad, float speed
     if (m_extTimeSrc && m_extTimeSrc->isEnabled())
     {
         qint64 displayedTime = getCurrentTime();
-        if (speed != 0  && (quint64)displayedTime != AV_NOPTS_VALUE)
+        if (speed != 0  && displayedTime != AV_NOPTS_VALUE)
         {
             bool firstWait = true;
             QTime sleepTimer;
@@ -1793,10 +1793,10 @@ qint64 QnCamDisplay::getDisplayedMin() const
     for (int i = 0; i < CL_MAX_CHANNELS && m_display[i]; ++i)
     {
         qint64 val = m_display[i]->getTimestampOfNextFrameToRender();
-        if ((quint64)val == AV_NOPTS_VALUE)
+        if (val == AV_NOPTS_VALUE)
             continue;
 
-        if ((quint64)rez == AV_NOPTS_VALUE)
+        if (rez == AV_NOPTS_VALUE)
             rez = val;
         else
             rez = qMin(rez, val);
@@ -1826,7 +1826,7 @@ qint64 QnCamDisplay::getMinReverseTime() const
     qint64 rez = m_nextReverseTime[0];
     for (int i = 1; i < CL_MAX_CHANNELS && m_display[i]; ++i)
     {
-        if ((quint64)m_nextReverseTime[i] != AV_NOPTS_VALUE && m_nextReverseTime[i] < rez)
+        if (m_nextReverseTime[i] != AV_NOPTS_VALUE && m_nextReverseTime[i] < rez)
             rez = m_nextReverseTime[i];
     }
     return rez;
@@ -1850,7 +1850,7 @@ qint64 QnCamDisplay::getNextTime() const
 
     qint64 lastTime = m_display[0]->getTimestampOfNextFrameToRender();
 
-    if ((quint64)rez != AV_NOPTS_VALUE)
+    if (rez != AV_NOPTS_VALUE)
         return m_speed < 0 ? qMin(rez, lastTime) : qMax(rez, lastTime);
     else
         return lastTime;
